+++ packages/okta-auth-js/test/spec/browserStorage.js
@@ -9,56 +9,69 @@ describe('browserStorage', () => {
   let originalLocation;
 
   beforeEach(() => {
-    originalLocalStorage = window.localStorage;
-    originalSessionStorage = window.sessionStorage;
-    originalLocation = window.location;
+    originalLocalStorage = global.window.localStorage;
+    originalSessionStorage = global.window.sessionStorage;
+    originalLocation = global.window.location;
   });
 
   afterEach(() => {
-    window.localStorage = originalLocalStorage;
-    window.sessionStorage = originalSessionStorage;
-    window.location = originalLocation;
+    global.window.localStorage = originalLocalStorage;
+    global.window.sessionStorage = originalSessionStorage;
+    global.window.location = originalLocation;
+    storageBuilder.mockClear();
   });
 
   it('can return localStorage', () => {
-    expect(window.localStorage).toBeDefined();
-    expect(browserStorage.getLocalStorage()).toBe(global.localStorage);
+    expect(global.window.localStorage).toBeDefined();
+    expect(browserStorage.getLocalStorage()).toBe(global.window.localStorage);
+    expect(browserStorage.getLocalStorage()).not.toBe(global.window.sessionStorage);
   });
 
   it('can return sessionStorage', () => {
-    expect(window.sessionStorage).toBeDefined();
-    expect(browserStorage.getSessionStorage()).toBe(global.sessionStorage);
+    expect(global.window.sessionStorage).toBeDefined();
+    expect(browserStorage.getSessionStorage()).toBe(global.window.sessionStorage);
   });
 
   describe('browserHasLocalStorage', () => {
+
     it('returns true if storage exists and passes test', () => {
       expect(browserStorage.browserHasLocalStorage()).toBe(true);
     });
+
     it('returns false if localStorage does not exist', () => {
-      delete window.localStorage;
+      delete global.window.localStorage;
       expect(browserStorage.browserHasLocalStorage()).toBe(false);
     });
+
     it('returns false if testStorage() returns false', () => {
       jest.spyOn(browserStorage, 'testStorage').mockReturnValue(false);
       expect(browserStorage.browserHasLocalStorage()).toBe(false);
     });
+
   });
 
   describe('browserHasSessionStorage', () => {
+
     it('returns true if storage exists and passes test', () => {
       expect(browserStorage.browserHasSessionStorage()).toBe(true);
+      expect(browserStorage.getSessionStorage()).toBe(global.window.sessionStorage);
+      expect(browserStorage.getSessionStorage()).not.toBe(global.window.localStorage);
     });
+
     it('returns false if sessionStorage does not exist', () => {
       delete window.sessionStorage;
       expect(browserStorage.browserHasSessionStorage()).toBe(false);
     });
+
     it('returns false if testStorage() returns false', () => {
       jest.spyOn(browserStorage, 'testStorage').mockReturnValue(false);
       expect(browserStorage.browserHasSessionStorage()).toBe(false);
     });
+
   });
 
   describe('testStorage', () => {
+
     it('returns true if no exception is thrown', () => {
       const fakeStorage = {
         removeItem: jest.fn(),
@@ -68,6 +81,7 @@ describe('browserStorage', () => {
       expect(fakeStorage.setItem).toHaveBeenCalledWith('okta-test-storage', 'okta-test-storage');
       expect(fakeStorage.removeItem).toHaveBeenCalledWith('okta-test-storage');
     });
+
     it('returns false if an exception is thrown on removeItem', () => {
       const fakeStorage = {
         removeItem: jest.fn().mockImplementation(() => {
@@ -77,6 +91,7 @@ describe('browserStorage', () => {
       }
       expect(browserStorage.testStorage(fakeStorage)).toBe(false);
     });
+
     it('returns false if an exception is thrown on setItem', () => {
       const fakeStorage = {
         removeItem: jest.fn(),
@@ -86,18 +101,28 @@ describe('browserStorage', () => {
       }
       expect(browserStorage.testStorage(fakeStorage)).toBe(false);
     });
+
   });
 
   describe('getPKCEStorage', () => {
-    it('Uses localStorage by default', () => {
+
+    it('Uses sessionStorage by default', () => {
       browserStorage.getPKCEStorage();
-      expect(storageBuilder).toHaveBeenCalledWith(global.window.localStorage, 'okta-pkce-storage');
+      expect(storageBuilder).toHaveBeenCalledTimes(1);
+      // .toHaveBeenCalledWith doesn't do a strict comparison, so an empty localStorage reads the same as an empty sessionStorage
+      expect(storageBuilder.mock.calls[0][0]).toBe(global.window.sessionStorage);
+      expect(storageBuilder.mock.calls[0][1]).toBe('okta-pkce-storage');
     });
-    it('Uses sessionStorage if localStorage is not available', () => {
-      delete global.window.localStorage;
+
+    it('Uses localStorage if sessionStorage is not available', () => {
+      delete global.window.sessionStorage;
       browserStorage.getPKCEStorage();
-      expect(storageBuilder).toHaveBeenCalledWith(global.window.sessionStorage, 'okta-pkce-storage');
+      expect(storageBuilder).toHaveBeenCalledTimes(1);
+      // .toHaveBeenCalledWith doesn't do a strict comparison, so an empty localStorage reads the same as an empty sessionStorage
+      expect(storageBuilder.mock.calls[0][0]).toBe(global.window.localStorage);
+      expect(storageBuilder.mock.calls[0][1]).toBe('okta-pkce-storage');
     });
+
     it('Uses cookie storage if localStorage and sessionStorage are not available', () => {
       delete global.window.localStorage;
       delete global.window.sessionStorage;
@@ -108,18 +133,28 @@ describe('browserStorage', () => {
       expect(storageBuilder).toHaveBeenCalledWith(fakeStorage, 'okta-pkce-storage');
       expect(browserStorage.getCookieStorage).toHaveBeenCalledWith(opts);
     });
+
   });
 
   describe('getHttpCache', () => {
+
     it('Uses localStorage by default', () => {
       browserStorage.getHttpCache();
-      expect(storageBuilder).toHaveBeenCalledWith(global.window.localStorage, 'okta-cache-storage');
+      expect(storageBuilder).toHaveBeenCalledTimes(1);
+      // .toHaveBeenCalledWith doesn't do a strict comparison, so an empty localStorage reads the same as an empty sessionStorage
+      expect(storageBuilder.mock.calls[0][0]).toBe(global.window.localStorage);
+      expect(storageBuilder.mock.calls[0][1]).toBe('okta-cache-storage');
     });
+
     it('Uses sessionStorage if localStorage is not available', () => {
       delete global.window.localStorage;
       browserStorage.getHttpCache();
-      expect(storageBuilder).toHaveBeenCalledWith(global.window.sessionStorage, 'okta-cache-storage');
+      expect(storageBuilder).toHaveBeenCalledTimes(1);
+      // .toHaveBeenCalledWith doesn't do a strict comparison, so an empty localStorage reads the same as an empty sessionStorage
+      expect(storageBuilder.mock.calls[0][0]).toBe(global.window.sessionStorage);
+      expect(storageBuilder.mock.calls[0][1]).toBe('okta-cache-storage');
     });
+
     it('Uses cookie storage if localStorage and sessionStorage are not available', () => {
       delete global.window.localStorage;
       delete global.window.sessionStorage;
@@ -130,9 +165,11 @@ describe('browserStorage', () => {
       expect(storageBuilder).toHaveBeenCalledWith(fakeStorage, 'okta-cache-storage');
       expect(browserStorage.getCookieStorage).toHaveBeenCalledWith(opts);
     });
+
   });
 
   describe('getCookieStorage', () => {
+    
     it('requires an options object', () => {
       const fn = function() {
         browserStorage.getCookieStorage();
@@ -181,15 +218,19 @@ describe('browserStorage', () => {
         sameSite: 'strictly fakey'
       });
     })
+    
   });
 
   describe('getInMemoryStorage', () => {
+
     it('can set and retrieve a value from memory', () => {
       const storage = browserStorage.getInMemoryStorage();
       const key = 'fake-key';
       const val = { fakeValue: true };
       storage.setItem(key, val);
       expect(storage.getItem(key)).toBe(val);
-    })
-  })
+    });
+
+  });
+
 });